✨ Add improved logger
